+++ back/src/main/java/zavrsni/rad/request/ArticleCreateRequest.java
@@ -4,12 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class ArticleCreateRequest {
     private String name;
     private Long amount;
-    private Long price;
+    private double price;
     private String supplier;
 }
